move qe-app-registry catalog source creation to new module
